Split _subprocess_execution.py along its stated seams (#117)
`cuprum/_subprocess_execution.py` was 510 lines, carrying both a
`# pylint: disable=too-many-lines` pragma and a `# TODO: refactor into
smaller submodules` — the violation was suppressed rather than fixed. The
module bundled stdin writing, stream-consumer spawning, timeout handling,
and the top-level runner.

Split along the seams the TODO already named:

- `cuprum/_subprocess_stdin.py` — `_emit_stdin_error`, `_write_stdin`,
  `_spawn_stdin_writer`, with the `cuprum.stdin` logger now living in the
  module whose name matches it.
- `cuprum/_subprocess_timeout.py` — the timeout dataclasses and errors,
  `_handle_subprocess_timeout` / `_raise_timeout_expired` /
  `_handle_stream_timeout`, the exit-event helpers they share with the
  normal completion path, and a single `_require_timeout` guard that
  centralises the duplicated "TimeoutError without a configured timeout"
  check.
- `cuprum/_subprocess_execution.py` keeps the runner:
  `_execute_subprocess`, `_run_subprocess_with_streams`,
  `_spawn_subprocess`, and the stream-consumer wiring.

Remove the `too-many-lines` pragma and the TODO (all three modules are
now 103-274 lines), and drop the redundant `_resolve_timeout` re-export
from `__all__` — `cuprum/sh.py` imports it from its definition site
`cuprum._subprocess_context`. `test_observe` patches `_write_stdin` at
its new home. The wheel-build snapshot reflects the new file list.
